+++ front/pages/w/[wId]/assistant/new.tsx
@@ -24,6 +24,7 @@ import { useRouter } from "next/router";
 import type { ReactElement } from "react";
 import { useContext, useEffect, useState } from "react";
 
+import { ReachedLimitPopup } from "@app/components/app/ReachedLimitPopup";
 import { AssistantDetails } from "@app/components/assistant/AssistantDetails";
 import type { ConversationLayoutProps } from "@app/components/assistant/conversation/ConversationLayout";
 import ConversationLayout from "@app/components/assistant/conversation/ConversationLayout";
@@ -446,72 +447,13 @@ export default function AssistantNew({
         onSubmit={handleSubmit}
         conversationId={null}
       />
-      <LimitReachedPopup
+      <ReachedLimitPopup
         isOpened={planLimitReached}
         onClose={() => setPlanLimitReached(false)}
         subscription={subscription}
         workspaceId={owner.sId}
       />
     </>
-  );
-}
-
-function getUpsellDialogDetailsForFreeTrial() {
-  return {
-    title: "You’ve reach the Free Trial daily messages limit",
-    validateLabel: "Skip trial & upgrade now",
-    children: (
-      <p className="text-sm font-normal text-element-800">
-        Come back tomorrow for a fresh start or{" "}
-        <span className="font-bold">skip the trial and start paying now.</span>
-      </p>
-    ),
-  };
-}
-
-function getUpsellDialogDetailsForFreePlan() {
-  return {
-    title: "You’ve reach the messages limit",
-    validateLabel: "Check Dust plans",
-    children: (
-      <p className="text-sm font-normal text-element-800">
-        Looks like you've used up all your messages. Check out our paid plans to
-        get unlimited messages.
-      </p>
-    ),
-  };
-}
-
-export function LimitReachedPopup({
-  isOpened,
-  onClose,
-  subscription,
-  workspaceId,
-}: {
-  isOpened: boolean;
-  onClose: () => void;
-  subscription: SubscriptionType;
-  workspaceId: string;
-}) {
-  const router = useRouter();
-
-  const { children, title, validateLabel } = isTrial(subscription)
-    ? getUpsellDialogDetailsForFreeTrial()
-    : getUpsellDialogDetailsForFreePlan();
-
-  return (
-    <Dialog
-      isOpen={isOpened}
-      title={title}
-      onValidate={() => {
-        void router.push(`/w/${workspaceId}/subscription`);
-      }}
-      onCancel={() => onClose()}
-      cancelLabel="Close"
-      validateLabel={validateLabel}
-    >
-      {children}
-    </Dialog>
   );
 }
